+++ src/action/env.rs
@@ -1,35 +1,35 @@
-use crate::{cli::EnvCmd as Cmd, Context, Ctx, QuartzResult, StateField};
+use crate::{cli::EnvCmd as Cmd, Ctx, Env, QuartzResult, StateField};
 use colored::Colorize;
 
 pub fn cmd(ctx: &Ctx, command: Cmd) -> QuartzResult {
     match command {
         Cmd::Create { name } => create(ctx, name),
         Cmd::Cp { src, dest } => cp(ctx, src, dest)?,
-        Cmd::Use { context } => switch(ctx, context)?,
+        Cmd::Use { env } => switch(ctx, env)?,
         Cmd::Ls => ls(ctx),
-        Cmd::Rm { context } => rm(ctx, context),
+        Cmd::Rm { env } => rm(ctx, env),
     };
 
     Ok(())
 }
 
 pub fn create(ctx: &Ctx, name: String) {
-    let context = Context::new(&name);
+    let env = Env::new(&name);
 
-    if context.exists(ctx) {
-        panic!("a context named {} already exists", name.red());
+    if env.exists(ctx) {
+        panic!("a environment named {} already exists", name.red());
     }
 
-    if context.write(ctx).is_err() {
-        panic!("failed to create {} context", name);
+    if env.write(ctx).is_err() {
+        panic!("failed to create {} environment", name);
     }
 }
 
 pub fn cp(ctx: &Ctx, src: String, dest: String) -> QuartzResult {
-    let src = Context::parse(ctx, &src).unwrap_or_else(|_| {
-        panic!("no {} context found", &src);
+    let src = Env::parse(ctx, &src).unwrap_or_else(|_| {
+        panic!("no {} environment found", &src);
     });
-    let mut dest = Context::parse(ctx, &dest).unwrap_or(Context::new(&dest));
+    let mut dest = Env::parse(ctx, &dest).unwrap_or(Env::new(&dest));
 
     for (key, value) in src.variables.iter() {
         dest.variables.insert(key.to_string(), value.to_string());
@@ -44,53 +44,53 @@ pub fn cp(ctx: &Ctx, src: String, dest: String) -> QuartzResult {
     Ok(())
 }
 
-pub fn switch(ctx: &Ctx, context: String) -> QuartzResult {
-    let context = Context::new(&context);
+pub fn switch(ctx: &Ctx, env: String) -> QuartzResult {
+    let env = Env::new(&env);
 
-    if !context.exists(ctx) {
-        panic!("context {} does not exist", context.name.red());
+    if !env.exists(ctx) {
+        panic!("environment {} does not exist", env.name.red());
     }
 
-    if let Ok(()) = StateField::Context.set(ctx, &context.name) {
-        println!("Switched to {} context", context.name.green());
+    if let Ok(()) = StateField::Env.set(ctx, &env.name) {
+        println!("Switched to {} environment", env.name.green());
     } else {
-        panic!("failed to switch to {} context", context.name.red());
+        panic!("failed to switch to {} environment", env.name.red());
     }
 
     Ok(())
 }
 
 pub fn ls(ctx: &Ctx) {
-    if let Ok(entries) = std::fs::read_dir(ctx.path().join("contexts")) {
+    if let Ok(entries) = std::fs::read_dir(ctx.path().join("env")) {
         for entry in entries {
             let bytes = entry.unwrap().file_name();
-            let context_name = bytes.to_str().unwrap();
+            let env_name = bytes.to_str().unwrap();
 
             let state = ctx
                 .state
-                .get(ctx, StateField::Context)
+                .get(ctx, StateField::Env)
                 .unwrap_or(String::from("default"));
 
-            if state == context_name {
-                println!("* {}", context_name.green());
+            if state == env_name {
+                println!("* {}", env_name.green());
             } else {
-                println!("  {}", context_name);
+                println!("  {}", env_name);
             }
         }
     }
 }
 
-pub fn rm(ctx: &Ctx, context: String) {
-    let context = Context::new(&context);
+pub fn rm(ctx: &Ctx, env: String) {
+    let env = Env::new(&env);
 
-    if !context.exists(ctx) {
-        panic!("context {} does not exist", context.name.red());
+    if !env.exists(ctx) {
+        panic!("environment {} does not exist", env.name.red());
     }
 
-    if std::fs::remove_dir_all(context.dir(ctx)).is_ok() {
-        println!("Deleted {} context", context.name.green());
+    if std::fs::remove_dir_all(env.dir(ctx)).is_ok() {
+        println!("Deleted {} environment", env.name.green());
     } else {
-        panic!("failed to delete {} context", context.name.red());
+        panic!("failed to delete {} environment", env.name.red());
     }
 }
 
@@ -98,7 +98,7 @@ pub fn print(ctx: &Ctx) {
     println!(
         "{}",
         ctx.state
-            .get(ctx, StateField::Context)
+            .get(ctx, StateField::Env)
             .unwrap_or("default".into())
     );
 }