+++ tests/integration_test_cases/env.rs
@@ -1,10 +1,10 @@
 use crate::utils::*;
 
 #[test]
-fn it_starts_with_default_context() -> TestResult {
+fn it_starts_with_default_env() -> TestResult {
     let quartz = Quartz::preset_empty_project()?;
 
-    let output = quartz.cmd(&["show", "ctx"])?;
+    let output = quartz.cmd(&["show", "env"])?;
 
     assert!(output.status.success(), "{}", output.stderr);
     assert_eq!(output.stdout.trim(), "default");
@@ -13,38 +13,38 @@ fn it_starts_with_default_context() -> TestResult {
 }
 
 #[test]
-fn it_can_create_context() -> TestResult {
+fn it_can_create_env() -> TestResult {
     let quartz = Quartz::preset_empty_project()?;
 
-    let output = quartz.cmd(&["ctx", "create", "example"])?;
-    let list = quartz.cmd(&["ctx", "list"])?;
+    let output = quartz.cmd(&["env", "create", "example"])?;
+    let list = quartz.cmd(&["env", "list"])?;
 
     assert!(output.status.success(), "{}", output.stderr);
     assert!(
         list.stdout.contains("example"),
-        "new context did not show on list"
+        "new env did not show on list"
     );
 
     Ok(())
 }
 
 #[test]
-fn it_can_switch_between_contexts() -> TestResult {
+fn it_can_switch_between_envs() -> TestResult {
     let quartz = Quartz::preset_empty_project()?;
 
-    quartz.cmd(&["ctx", "create", "example"])?;
-    quartz.cmd(&["ctx", "create", "lorem"])?;
+    quartz.cmd(&["env", "create", "example"])?;
+    quartz.cmd(&["env", "create", "lorem"])?;
 
     let assert_use = |name: &str| -> TestResult {
-        let output = quartz.cmd(&["ctx", "use", name])?;
-        let status = quartz.cmd(&["show", "ctx"])?;
+        let output = quartz.cmd(&["env", "use", name])?;
+        let status = quartz.cmd(&["show", "env"])?;
         assert!(output.status.success(), "{}", output.stderr);
         assert_eq!(status.stdout.trim(), name, "did not switch to {}", name);
 
         Ok(())
     };
 
-    assert_use("lorem")?; // Using the newly created context
+    assert_use("lorem")?; // Using the newly created env
     assert_use("example")?; // Switch to another one
     assert_use("lorem")?; // Go back to previously selected
     assert_use("default")?; // Going to default
@@ -53,48 +53,48 @@ fn it_can_switch_between_contexts() -> TestResult {
 }
 
 #[test]
-fn it_can_not_use_unexistent_context() -> TestResult {
+fn it_can_not_use_unexistent_env() -> TestResult {
     let quartz = Quartz::preset_empty_project()?;
 
-    let output = quartz.cmd(&["ctx", "use", "idontexist"])?;
+    let output = quartz.cmd(&["env", "use", "idontexist"])?;
     assert!(
         !output.status.success(),
         "{}",
-        "it allowed using unexistent context"
+        "it allowed using unexistent env"
     );
 
-    let status = quartz.cmd(&["show", "ctx"])?;
+    let status = quartz.cmd(&["show", "env"])?;
     assert_ne!(
         status.stdout.trim(),
         "idontexist",
-        "it is using unexistent context"
+        "it is using unexistent env"
     );
 
     Ok(())
 }
 
 #[test]
-fn it_can_remove_context() -> TestResult {
+fn it_can_remove_env() -> TestResult {
     let quartz = Quartz::preset_empty_project()?;
 
-    quartz.cmd(&["context", "create", "example"])?;
+    quartz.cmd(&["env", "create", "example"])?;
 
-    let output = quartz.cmd(&["context", "remove", "example"])?;
+    let output = quartz.cmd(&["env", "remove", "example"])?;
     assert!(output.status.success(), "{}", output.stderr);
-    let list = quartz.cmd(&["context", "list"])?;
+    let list = quartz.cmd(&["env", "list"])?;
     assert!(
         !list.stdout.contains("example"),
-        "example context is still listed"
+        "example env is still listed"
     );
 
     Ok(())
 }
 
 #[test]
-fn it_cannot_remove_unexistent_context() -> TestResult {
+fn it_cannot_remove_unexistent_env() -> TestResult {
     let quartz = Quartz::preset_empty_project()?;
 
-    let output = quartz.cmd(&["context", "remove", "example"])?;
+    let output = quartz.cmd(&["env", "remove", "example"])?;
     assert!(!output.status.success(), "did not exit with error");
 
     Ok(())
@@ -104,9 +104,9 @@ fn it_cannot_remove_unexistent_context() -> TestResult {
 fn it_cannot_create_duplicate() -> TestResult {
     let quartz = Quartz::preset_empty_project()?;
 
-    quartz.cmd(&["context", "create", "example"])?;
-    quartz.cmd(&["context", "create", "lorem"])?;
-    let output = quartz.cmd(&["context", "create", "example"])?;
+    quartz.cmd(&["env", "create", "example"])?;
+    quartz.cmd(&["env", "create", "lorem"])?;
+    let output = quartz.cmd(&["env", "create", "example"])?;
 
     assert!(
         !output.status.success(),