+++ tests/integration_test_cases/mod.rs
@@ -1,8 +1,8 @@
 pub mod body;
 pub mod config;
-pub mod context;
-pub mod context_variable;
 pub mod endpoints;
+pub mod env;
 pub mod headers;
 pub mod init;
 pub mod query;
+pub mod var;