+++ tests/utils/mod.rs
@@ -67,10 +67,10 @@ impl Quartz {
         Ok(quartz)
     }
 
-    pub fn preset_using_default_context() -> Result<Self, std::io::Error> {
+    pub fn preset_using_default_env() -> Result<Self, std::io::Error> {
         let quartz = Quartz::preset_empty_project()?;
 
-        quartz.cmd(&["context", "use", "default"])?;
+        quartz.cmd(&["env", "use", "default"])?;
 
         Ok(quartz)
     }